sprint 1
i have amade a log in screen that would let user to log in and sign up. i have connected my app to Firebase and created options to log in with existing ccount or create a new account and log in.  Also i have added a function that keeps user logged in when the app is closed but i had to comment it as it makes it difficult to test